+++ LocalPayment/src/main/java/com/braintreepayments/api/localpayment/LocalPaymentAuthRequestParams.kt
@@ -10,12 +10,11 @@ import com.braintreepayments.api.BrowserSwitchOptions
  * @property approvalUrl used for payment approval.
  * @property paymentId of the local payment after creation.
  */
-
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 data class LocalPaymentAuthRequestParams internal constructor(
     val request: LocalPaymentRequest,
     val approvalUrl: String,
     val paymentId: String
 ) {
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
     var browserSwitchOptions: BrowserSwitchOptions? = null
 }