+++ LocalPayment/src/main/java/com/braintreepayments/api/localpayment/LocalPaymentClient.kt
@@ -56,21 +56,14 @@ class LocalPaymentClient @VisibleForTesting internal constructor(
      * @param callback [LocalPaymentAuthCallback]
      */
     fun createPaymentAuthRequest(
-        request: LocalPaymentRequest?,
-        callback: LocalPaymentAuthCallback?
+        request: LocalPaymentRequest,
+        callback: LocalPaymentAuthCallback
     ) {
         analyticsParamRepository.resetSessionId()
         braintreeClient.sendAnalyticsEvent(LocalPaymentAnalytics.PAYMENT_STARTED)
 
         var exception: Exception? = null
-
-        if (callback == null) {
-            throw IllegalArgumentException("A LocalPaymentAuthRequestCallback is required.")
-        }
-
-        if (request == null) {
-            exception = BraintreeException("A LocalPaymentRequest is required.")
-        } else if (request.paymentType == null || request.amount == null) {
+        if (request.paymentType == null || request.amount == null) {
             exception = BraintreeException(
                 "LocalPaymentRequest is invalid, paymentType and amount are required."
             )
@@ -174,7 +167,7 @@ class LocalPaymentClient @VisibleForTesting internal constructor(
         callback: LocalPaymentTokenizeCallback
     ) {
         val browserSwitchResult: BrowserSwitchFinalResult.Success = localPaymentAuthResult
-            .paymentAuthInfo.browserSwitchResultInfo
+            .browserSwitchSuccess
 
         val metadata: JSONObject? = browserSwitchResult.requestMetadata
         val merchantAccountId = Json.optString(metadata, "merchant-account-id", null)