+++ SEPADirectDebit/src/main/java/com/braintreepayments/api/sepadirectdebit/SEPADirectDebitPaymentAuthRequestParams.kt
@@ -1,13 +1,12 @@
 package com.braintreepayments.api.sepadirectdebit
 
+import androidx.annotation.RestrictTo
 import com.braintreepayments.api.BrowserSwitchOptions
 
 /**
  * Returned via the [SEPADirectDebitPaymentAuthRequestCallback] after calling
  * [SEPADirectDebitClient.createPaymentAuthRequest].
  *
- * Inspect the [SEPADirectDebitNonce] property to determine if tokenization is complete, or
- * if you must continue the SEPA mandate web flow via
- * [SEPADirectDebitLauncher.launch]
  */
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 class SEPADirectDebitPaymentAuthRequestParams internal constructor(val browserSwitchOptions: BrowserSwitchOptions)