+++ Venmo/src/main/java/com/braintreepayments/api/venmo/VenmoPaymentAuthRequestParams.kt
@@ -1,12 +1,12 @@
 package com.braintreepayments.api.venmo
 
+import androidx.annotation.RestrictTo
 import com.braintreepayments.api.BrowserSwitchOptions
 
 /**
  * Used to request Venmo authentication via [VenmoLauncher.launch] )}
  */
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 data class VenmoPaymentAuthRequestParams internal constructor(
-    // TODO: this should be internal and restricted to library scope but can't be until
-    //  VenmoClientUnitTest is converted to Kotlin
     val browserSwitchOptions: BrowserSwitchOptions
 )